phase-4.2-realtime-optimization/requirements.txt
# Phase 4.2: Real-time Optimization Dependencies

# Core dependencies
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
psutil>=5.9.0

# Data processing and ML
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.1.0

# Embedding generation
sentence-transformers>=2.2.0
torch>=1.12.0
transformers>=4.21.0
openai>=1.0.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=0.19.0
pydantic>=1.10.0

# Logging and monitoring
structlog>=22.1.0
pythonjsonlogger>=2.0.0

# Testing
pytest>=7.1.0
pytest-asyncio>=0.20.0
pytest-mock>=3.8.0
pytest-cov>=4.0.0

# Development tools
black>=22.6.0
flake8>=5.0.0
mypy>=0.971
isort>=5.10.0

# Performance monitoring
matplotlib>=3.5.0
seaborn>=0.11.0
